# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SindÂncora API",
        "description": "API REST do SindÂncora — SaaS de Gestão Condominial",
        "contact": {
            "email": "suporte@sindancora.com.br"
        },
        "license": {
            "name": "Proprietário"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Base"
        }
    ],
    "paths": {
...
        "/v1/app/reservations": {
            "get": {
                "tags": [
                    "App"
                ],
                "summary": "Listar reservas (filtros: status, common_area_id, condominium_id)",
                "operationId": "appReservationsIndex",
                "responses": {
                    "200": {
                        "description": "Lista paginada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/app/reservations/{id}": {
            "get": {
                "tags": [
                    "App"
                ],
                "summary": "Detalhe da reserva",
                "operationId": "appReservationsShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reserva"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/app/reservations/{id}/approve": {
            "post": {
                "tags": [
                    "App"
                ],
                "summary": "Aprovar reserva pendente",
                "operationId": "appReservationsApprove",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reserva aprovada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/app/reservations/{id}/reject": {
            "post": {
                "tags": [
                    "App"
                ],
                "summary": "Rejeitar reserva pendente (reason opcional)",
                "operationId": "appReservationsReject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reserva rejeitada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/app/reservations/{id}/cancel": {
            "post": {
                "tags": [
                    "App"
                ],
                "summary": "Cancelar reserva (reason opcional)",
                "operationId": "appReservationsCancel",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reserva cancelada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "apiKey": {
                "type": "http",
                "description": "API Key do tenant no formato sk_live_... enviada como Bearer token.",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}